Missing changes from merge
